New 'loading' feature for overlay layers. It's visible as a spinning icon in the right panel, next to each layer being loaded.
There is an important limitation:
We want to show the 'loading' icon everytime new tiles
are being loaded, but we can't know using the
Google Maps API if new tiles need to be loaded.
We can detect events like the user dragging the map around,
but that doesn't guarantee that new tiles are going to be
requested.

e.g. Consider a tile bigger than the viewport. If you move
the map just a little, but not enough to go out of the bounds
of the tile, no new tiles are needed.
So user events are not good indicators of tile loading.
We could overcome this by migrating to OpenLayers. ;)