General database optimizations. ===============================
This update adds several indexes to the Guifi module that have been proved to increase the
performance in some queries. In some of them, they have gone from taking several seconds
to being instant (i.e. tens of seconds vs milliseconds).

Here I list some of the improved areas:
 - Searching for nodes between an area (latitude/longitude) has been improved from O(n^2) to
O(n), where 'n' = total node count. This can be further improved up to O(log n) if we make use
of MySQL Spatial Extensions (installed by default in MySQL) or PostgrSQL + PostGIS extensions.
Working on it.
 - Listing the nodes of a Guifi zone has been improved. This is done in every zone page, and
probably in others too. It is O(log n) now.
 - Listing recent nodes improved.
 - Improved generation of CNML.
 - Improved generation of statistics: e.g. total node count for every status, total link
length, etc
 - Improved growthmap generation.
 - Improved traceroute page.
 - Improved listing of services by zone.
 - Improved asking for the childs of a zone.

And probably the changes affect (positively) other areas of the web too.
Nevertheless the biggest resource hog is Drupal itself.
It is recommended to run ANALYZE TABLE and OPTIMIZE TABLE for every table in the DB once a
week for optimum performance. There is a DB Maintenance module for Drupal that can do this
automatically.