Added 'dynamic nodes' feature.
Now the javascript maps can query an area for nodes and display them dynamically, with
different styles depending of various factors (e.g. the bigger the supernode, the bigger the
marker).
Clicking on them displays basic information, such as name of the node, status, etc. and
clicking in the name opens the node page.

Setting the markers invisible in the source code allows to keep the clickable area, but not
render the markers. The visual part can come from another layer.

This feature is disabled by default (it can be slow due to high CPU usage), but can be
enabled in the right panel. There is a hard limit of 300 nodes displayed at once in an area
(configurable in the source code).